+++ examples/futures/trade/cancel_plan_order.py
@@ -16,7 +16,34 @@ futuresAPI = APIContract(api_key="Your_Api_Key",
 try:
     response = futuresAPI.post_cancel_plan_order(
         contract_symbol="ETHUSDT",
+    )[0]
+    logger.info(response)
+except APIException as error:
+    logger.error(
+        "Found error. status: {}, error message: {}".format(
+            error.status_code, error.response
+        )
+    )
+
+try:
+    response = futuresAPI.post_cancel_plan_order(
+        contract_symbol="ETHUSDT",
         order_id='12312312',
+    )[0]
+    logger.info(response)
+except APIException as error:
+    logger.error(
+        "Found error. status: {}, error message: {}".format(
+            error.status_code, error.response
+        )
+    )
+
+
+
+try:
+    response = futuresAPI.post_cancel_plan_order(
+        contract_symbol="ETHUSDT",
+        client_order_id='12312312',
     )[0]
     logger.info(response)
 except APIException as error: