+++ examples/futures/trade/modify_plan_order.py
@@ -14,7 +14,14 @@ futuresAPI = APIContract(api_key="Your_Api_Key",
                          logger=logger)
 
 try:
-    response = futuresAPI.get_current_plan_order()[0]
+    response = futuresAPI.post_modify_plan_order(
+            contract_symbol='BTCUSDT',
+            client_order_id="12314323424",
+            trigger_price="2000",
+            executive_price="1450",
+            price_type=1,
+            type="limit"
+        )[0]
     logger.info(response)
 except APIException as error:
     logger.error(
@@ -23,17 +30,3 @@ except APIException as error:
         )
     )
 
-try:
-    response = futuresAPI.get_current_plan_order(
-        contract_symbol='BTCUSDT',
-        type='limit',
-        limit=10,
-        plan_type="plan",
-    )[0]
-    logger.info(response)
-except APIException as error:
-    logger.error(
-        "Found error. status: {}, error message: {}".format(
-            error.status_code, error.response
-        )
-    )